FIREFLY-37 update per review comments.
